Move Map Lua API to separate file
