+++ Makefile
@@ -258,6 +258,7 @@ obj/light_data.o \
 obj/lua_api.o \
 obj/lua_api_camera.o \
 obj/lua_api_lens.o \
+obj/lua_api_map.o \
 obj/lua_api_player.o \
 obj/lua_api_room.o \
 obj/lua_api_things.o \
