+++ config/fxdata/lua/classes/Map.lua
@@ -4,10 +4,9 @@
 ---@field map_name string The name of the map
 ---@field map_number integer The map number
 ---@field campaign string The campaign/Mappack this map belongs to
----@field map_typ string The type of map (Campaign, Multiplay, Bonus, Moon, MapPack)
+---@field map_type string The type of map (Campaign, Multiplayer, Bonus, Moon, MapPack)
 ---@field width integer The width of the map in tiles
 ---@field height integer The height of the map in tiles
 ---@field default_texture integer|string The default texture for the map, id for custom textures, or string for built-in textures
----@field creature_pool table A table containing the creature pool for the map, with creature types as keys and their counts as values
-
+---@field creature_pool table<string, integer> A table containing the creature pool for the map, with creature types as keys and their counts as values
 if not Map then Map = {} end
